Make ENFIX an ordinary cell flag, remove SET/ENFIX
There was a strange and probably ultimately misguided idea in old Ren-C
to limit the spread of the enfix cell flag, so that if you assigned one
action to another you wouldn't get its enfixedness.

Offhand I don't remember what caused the particular obsession with making
the enfix flag so weird, that it wouldn't be copied across cells.  But
the idea was dropped and infix flags were made ordinary and transferable
a long while ago in the mainline.

It's easy enough to change it, so this changes it to be in alignment
just to minimize headaches going back and forth...and less weird
properties to debug.