Make inert fetch state byte vs. eval flag
One of the first thing the stepper executor did was check an eval
flag to see if the evaluator was in "non-evaluative" mode.  This seemed
like something that could just be done with a state byte, which uses
a sunk cost of a switch statement and frees up an evaluator flag.

However, the "restarting" of steppers would overwrite this state
byte.  So that called into question whether steppers should just auto
reset themselves when they finish an evaluation, and the restart could
just be changed to an assert.

This does those changes which are simplifying and improve performance.