Get rid of action executor dispatcher_base
This was in all cases equal to the baseline stack base.  The reason
that dispatcher_base had originally existed was because pending actions
in cascades were pushed to the stack previously, but that method has been
completely redone...and it is cleaner to not have accrued stack in a
level when running action dispatchers.