Pare down usages of rebJumps(), Bounce as void*
With the advent of rebDelegate() you can now cooperatively hand back
control to the trampoline to run a failure on your behalf, while your
native is still "on the stack".  This is cleaner than allowing longjmp()
or exceptions to cross your native code stack.  There are still cases
which do abrupt fail for the sake of ease of coding, but these should
be reviewed.

Also gives in and makes Bounce just a void*.  Attempts to make it a
class did not work because a class holding a single pointer did not
behave the same in some compilation scenarios as the void* itself.
And having it be Node* wasn't offering enough benefit to warrant the
conflicting definition with the API...plus having it be void* keeps
it checked in the C++ build from being used with the variadic API.