First pass at streamlining literal left lookback
The literal left lookback code is complex, and needs a going over.  This
first pass at it found an entire unused clause that was based on testing
for whether quoted parameters were evaluated--which they couldn't be.

This does a simplification that is hopefully a prelude to more, as
well as optimization to cache the parameter class of the first argument
of functions (vs. just whether it's quoted, since we now need to know
if it needs binding or not.)