Replace LOAD-VALUE with TRANSCODE:ONE
LOAD used to act strangely when you gave it a string that resolved to a
single value and would give you a single value back.  So you had to use
LOAD:ALL to get it to give you that value in a block.  This was not
uniform, and so LOAD-VALUE was established to enforce getting back one
and only one value.

There was some thought that this had value over TRANSCODE:ONE due to
the idea that this value might be bound in some way by LOAD.  But now
that LOAD is concieved as an operation that requires a header, then it
definitely no longer makes sense to consider it as LOAD...also an unbound
value is far more useful now than it was in the past.