Use Level's STATE byte for scanning mode
This gets rid of the separate field in the scan state for the scanning
mode, and uses the state byte...which can be initialized in the flags
when starting.

This also allows for whether we're doing an interstitial scan or not
to be reflected in a flag set up with the level, so it can be tested
as one flag instead of testing against one of three bytes.