StackValue(*) => OnStack(Value*), OnStack(Element*)
This lets the property of being on the stack apply to cell subclasses.

Being tracked in this way lets us count the number of outstanding stack
cells, and notice if that's non-zero if there are any PUSH() or DROP().
This protects against bugs that would only be visible in the case of
stack expansions (or in the future, contractions).