Renaming related to MOLD_STRUCT
Help convey that the binary flex in a mold struct holds UTF-8.

The modern executable does this by calling it "string" but strings are
not UTF-8 internally in the bootstrap executable.

Rather than calling it string or changing it to vague "buffer" this
accepts that the new and old executables will call it something
different, and a Mold_Buffer() macro can abstract it where it makes
sense for compatibility.