Don't put custom flag used-with type in flag
The bootstrap executable had extra checking to help you avoid checking
flags in the header for things that weren't applicable to that type.
Modern Ren-C saturates the header with other things (like a quote
level) so types don't try to exploit the cell header bits very much.
Given that the feature didn't get that much use the paranoid checks
are overkill, and just make the bootstrap executable unable to
easily copy/paste code with the new EXE.