Add primitive --{...}-- string scanning
This is an early test of scanning strings with dashed delimiters.
It is using the idea of making it possible to graft code onto the
old executable's scanner from the new one.  This way, as updates
are made to improve the new scanner they can be brought to bear
upon the bootstrap.

It also brings in the constraint that you can't put zero bytes in
strings (at least scanned ones, here) so tests have to be adjusted
to account for that.