Update bootstrap to --{...}--, HELP uses CSCAPE
This goes through and selectively converts {...} strings into
"..." strings, or uses some variation of -{...}- or --{...}--
The heuristics for picking what to use are evolving, but this is
kind of a first stylistic guess.

Because {...} was used in a very ugly way in the make.r's HELP,
this went ahead and remodeled that help on top of CSCAPE, and it
looks a lot better.