Get rid of MOLD:ALL and MOLD_FLAG_ALL
It had mostly been excised, but the trigger had not been pulled.

It's not fully clear how users would decide when to ask for MOLD vs
MOLD:ALL and it had never really shown effective use.  Red has not
implemented it, which supports the empirical observation that it
is too poorly defined to be useful.

This makes more room for looking into better ideas.