Get rid of MOLD:ALL and related code
This aligns with the main branch in getting rid of MOLD:ALL as the
:ALL refinement was not something that had been made coherent or that
people were able to wield effectively.

Also wipes out the idea of global settings that change the base or
number of digits made by core molding.

Throws in change that SAVE only saves whole blocks as well (the code
testing extended features of SAVE are all gone in this bootstrap EXE
in any case.)