New conventions for QUIT
This changes QUIT to always take a value.  The modern executable makes
an exception for the console version of quit, but it can do this specific
exception because it is based on definitional quit and creates a new quit
function for each module.  The bootstrap executable allows no argument on
any quit by taking the value endably.