Do not bind construction syntax scans
Construction syntax is going out, and this was causing crashes when
trying to scan before the lib context was created.  Remove it.

This also gets rid of the construction syntax instances of #[true]
and #[false] in the boot code that were added for ~true~ and ~false~.
That never worked, because it uses the symbol table to recognize the
constructed word.  (Hence why there is Startup_True_And_False)