Switch to FENCE!-based multi-return circling
Due to wanting to "circle" complex expressions that @ may not apply
to in multi-return slots, it has long been thought that FENCE! would
make a better way to circle returns.

The movement to where @var signals "reuse bindings, don't make a
new one" also points in this direction--so the weird quoting searched
for by LET can be gotten rid of.

This means @ had to be changed from "I want this value, but as is and
do not decay" (which it wasn't right for), so trash (~) is used for
that signal.  Realizing that BLANK! could not be used for "optional
nameless thing) since there is no :_ then that was switched to #,
which probably looks better anyway.