Improve Binder to use automatically clean up
Ren-C had stuck more or less with the original idea of the bind table
from R3-Alpha, which was a hash table based on symbols that would be
populated from an object to map the object symbols to what index it
was in that object.  This table would be populated by walking the
object and before a binding operation to add the indices, and then
cleared out manually by walking the object again when the binding
was done.

But the model has changed significantly to where symbols themselves
are GC-able nodes, which can have chains of binding information
linked onto them with tiny quick-to-allocate and quick-to-free stubs.
These stubs can be linked together and then freed all at once by
walking a singly linked list on binder shutdown.

This simplifies and speeds up several places in the code.