Eliminate data stack usage in collector
The original R3-Alpha had a specialized "Collect Buffer" which was
used when scanning arrays (usually for SET-WORD!s).  Rather than have
a special buffer for this purpose, Ren-C switched over to using the
data stack.

However, there's enough information to know what's been newly collected
latent in the Binder that the collector uses.  This takes a step by
removing the need for the data stack in collect operations.