Rethink SingleHeart as multiplexed separate enum type
Static analysis caught errors in the checking of a returned "leading blank"
flag without having first checked that a sequence was of the "just 2
elements, one of which is blank" variety.

Rather than fix just those cases, this rethinks the return result to give
back a new "SingleHeart" type that fuses together the heart and the
boolean of the leading blank, in a way that is completely separate from
the 0 state of not being a sequence of that form.

This is clarifying and prevents errors.  The new singleheart type is also
pushed so all of its values are out of range from Heart and Kind, so you
can't accidentally mix them up.