Flex as Stub subclass, flesh out more struct-xxx.h files
Clang static analyzer motivated wanting a separate Hashlist_Len() which
would not return 0, so that it wouldn't think there was division by 0.
That meant making HashList a separate type from Flex.

But those definitions needed a place to go, so struct-map.h seemed like
the right place to put it.  The struct-xxx.h files are included before
the %tmp-internals.h so they can be in core function prototypes.

So long as this was being done, it motivated making struct-xxx.h files
for everything else.  This also takes a step toward separating the
Stub and Flex types--which things have been moving toward over time.
Though they're still entwined a bit, and this doesn't resolve every
detail of that.

(While this is a large reorganization, there should be no changes in
functionality.)