Fix length of map calculation, map hash divide by zero check
Clang static analyzer wasn't certain that Flex_Used() wouldn't
return 0, so it warned of a possible division by zero when there
was Flex_Used(hashlist).

This makes a special Hashlist_Num_Slots() routine that lets Clang
static analyzer see that it can't return 0.

It also fixes a bug in the calculation of the length of maps,
that hadn't been updated to the new "zombie" unreadable cells
that are used (nulls and other antiforms no longer allowed in
map pairlists--unreadable is a better choice overall).