Remove anachronistic MAKE of TUPLE! from URL!
Historical URL! was an any-string type and had a position (vs. Ren-C
URL! which is immutable and has no position).  There was some code
that captured a parse position in a URL! (as a URL!) and then called
MAKE TUPLE! on it.  That code is long gone.

(The MAKE code is really just "lingering ideas" from tuple, all the
sequence creation is done with TO and AS at the moment.