Restyle Back_Scan_Utf8_Char() to return an error
The routine for decoding UTF-8 sequences has cases that error, and
previously that returned nullptr.  At least in part because returning
an error would involve an expensive allocation.

Clang static analyzer didn't like that, and pointed out a lot of cases
that weren't doing good error handling.

This preallocates errors that might be returned by the routine.  That
lets it keep working in cases like searching for UTF-8 TEXT! in a
binary BLOB! that's not all UTF-8...without allocating thousands of
errors as it goes along every byte in the search.