Add "Cooperative FAIL()"
The dangers of using exception-based abrupt fail are shifting the
implementation to use rebDelegate() instead of rebJumps() in the API.

Internally, it's also possible to cut down on the number of usages
of the throw/longjmp() abrupt failure, by doing a cooperative error
that needs rescue with `return FAIL()` as analog to `return RAISE()`

One of the advantages is that anywhere this construct is used in
lieu of the abrupt failure, you can actually run the system on
a platform that doesn't offer either longjmp() or throw.  But it's
generally cleaner, while still offering the same automatic cleanup
benefits.

It will take time before more of the system is designed around
cooperative failures, but this gets a good percentage of the cases.