Simplify TICK handling, add sys-trampoline.h
This adds a TICK macro that when ticks are counted evaluates to the
counted tick, and 0 when it is not.  That makes it easier to write code
that is agnostic about whether it's a DEBUG_COUNT_TICKS build or not.

If DEBUG_COUNT_TICKS then the tick starts out at 1.  This helps make
sure that 0 is a unique state indicating no DEBUG_COUNT_TICKS.

Also uses inttypes.h to properly printf() format uint64_t