BROWSE imported to LIB normally
Historically, BROWSE needed a definition in LIB to use for binding
purposes (inside something like HELP) and then that had to be
overwritten by the process library, because if there wasn't
something to bind to up front it would be unbound.

Among the many advantages of "pure virtual binding" is that you
can add the browse definition to lib later and it will still find
it, so this removes the anachronistic "browse stub".