Don't use UINT32_MAX for "UNLIMITED"
To make the "unlimited integer" concept more robust, the system was
changed to use pointers for limits that could be unlimited.  Then
unlimited was a null pointer.

One place that was missed was inside the code for append/insert/change
because instead of using UNLIMITED and being found, it used UINT32_MAX.
Code was changed to not use unsigned integers due to general
prescriptivism against them.