Revert CONSTRUCT to arity-1, allow MAKE on OBJECT!
Quite a while prior to the introduction of antiforms, there was a
question of how to represent unset variables in an object.  The
weird idea that was tried was to give objects a spec, and then
any values that were in the spec but not in the object body would
be assumed to be unset.

The idea that CONSTRUCT would be arity-2 and take one of these
"object specs" could also permit the addition of type information
to the fields...like a function spec.  For a time this was done
with HAS as an arity-1 version that had no spec.

An additional concept of the time was to get rid of the idea that
MAKE could take an instance of a type.  It only took datatypes,
so you couldn't say e.g. (make [a b c] "d e") and get the same
behavior as (make block! "d e")  This meant (make some-obj [...])
was off the table, pushing people to using the arity-2 CONSTRUCT
for that purpose.

This jumble of ideas wasn't really committed to, leading to an
emulation layer that implemented the old MAKE SOME-OBJ semantics.
While modern Ren-C preserves the idea that you can't say MAKE on
a BLOCK! instance, it did restore the idea of MAKE on SOME-OBJ.

Since CONSTRUCT went back to an arity-1 purpose, this brings the
bootstrap executable into sync with that, since it runs %sysobj.r.
And it removes the need for emulation by having MAKE implement
the old logic directly.