Replace instances of BLANK! with NULL
The boostrap executable has a "nearly modern" null state that can't
be put in blocks (at least not easily) and is falsey.  It is thus
better in most cases than blank (which is still falsey for the ability
to bootstrap the bootstrap executable...)

This turns the non-boostrap instances of blank over to be nulls.  It
should be more likely to raise errors in dodgy circumstances, so
since it's easy enough to do in a couple minutes it was worth it.