Rethink MAKE as a Generic
At one point, MAKE was changed into a function that used a table of
hooks for each built-in datatype, instead of a generic.  That was in
order to "purify" the generic machinery such that it never dispatched
just a datatype like OBJECT! to the generic handler for objects, but
always passed an instance...MAKE and TO were violating that.

However, this led to a weird shape for the hook, because the hook
still had to handle the case of "sometimes an instance, sometimes a
datatype".  Most instance cases were removed (e.g. you could no longer
say (make [a b c] "d e f") as a synonym for (make block! "d e f").
Only objects worked this way with (make object! [...]) as well as
a distinct meaning for (make some-obj [...])

This does a rethink of that, where it goes back to being a generic,
with the instance cases handled by the generic for that instance's
type, but the special subdispatch to the hooks done by the generic
handler for the datatype.  This helps remove the idea of passing an
instance to the make hook functions, and makes it generalizable while
still keeping the generics focused on instances of their type only.