Round to integer as default
ROUND was changed to be type preserving, but ROUND-ing a decimal to remove
its decimal point is likely the most common form of ROUND.  TO INTEGER!
is now intended to be preserving, and so it will no longer be the way
of rounding (nor should it, as it's not nuanced enough).

This makes rounding decimals default to producing integers.  You can use
a :TO parameter of 0.1 to get the old behavior of rounding to the nearest
integer but keeping a decimal point.