Give Sink "teeth" by corrupting variables
The original idea of Sink() was that it would automatically corrupt
the argument in the debug build, in order to expose cases where an
initialization wasn't performed.  But Erase_Cell() could not be used,
as it would corrupt the bit patterns...and there were some mechanical
problems related to the shared template for Need() and Sink().

This sorts out those problems, and uses just the plain "free bit" of
the cell to give a cheap method of making the cell non-readable but
still writable.

It also improves comments, makes the corruption more "interesting"
in terms of the spread of values assigned, and makes it so that the
static analysis builds do not use the sinking so as to avoid making
it look like the values were assigned when being analyzed.

Plus, it seems that "Initability" wasn't really being checked.  At
some point it looks like the Assert_Cell_Writable() checks were
dropped instead of just checking to make sure that it was either
that or erased.  Bring that check back...