Create Init() as alternative to Sink()
Sink() is mostly for code that branches.  If initialization code runs in
a straight line, then corrupting the thing you're about to initialize
doesn't really buy you very much, it just slows things down.

This makes a definition of Init() which can act like Sink() if you want,
but by default acts like Need()... so all it does is confirm that the
callsite can accommodate the cell subclass specified.