Clean up cell freshness/erasedness code
The ideas of a "fresh" vs. "erased" cell have evolved over time, and this
is a retrospective that puts things in the right order to that it coheres
better from top to bottom.

It standardizes the method for freshening cells while suppressing raised
errors better, and makes this freshening happen on every stack level
that returns a BOUNCE_THROWN vs. just the originating level.

Erasing is explained better in contrast to freshening, and Freshen_Cell()
is changed to not return a result, for a slight efficiency gain.