Rename allocation macros from Free()
When all-caps macros were phased out (for cases that don't have "weird"
effects, like being an LValue), the FREE() macro became Free().  But
this is now contentious with single-word leading-caps being a DataType.

Free_Memory() is slightly longer, but it's legible and doesn't look
like a datatype under the new rules.  Apply that pattern.