Fix valgrind memory leak, "potential" leaks
The debug build wasn't always doing a 'clean' shutdown as it was supposed
to, so it was making it look like there was a lot of potentially leaked
memory if using valgrind.  This switches to use a clean shutdown always
in the debug build.

(Clean shutdown could be controlled in release builds with an environment
variable or with a command line switch, if one wanted to test that the
release build didn't leak.)