Shut down GC correctly when GC is disabled
The system requires shutdown to make a legal state for startup, and
startup of the GC expected the GC disabled flag to be false.  But
commands like RECYCLE can disable it, and then you go through
shutdown and it needs to be cleared before another startup might
run... (we want to be able to shut down and start up again within
the same process).