Fix leak of uv_default_loop()
If you don't call uv_close_loop(), then a memory allocation of the default
event loop leaks in libuv.

In order to close the loop all the requests have to be closed, and uv_close()
on a timer is a request...you still have to run the loop before closing
the loop.