Fix CELL_FLAG_PROTECTED writability check, canon unreadable
It turned out that at some point, the read and writability checks had
been changed to operate on the node byte... which meant totally missing
the protected flag for the cell being checked.

This fixes that, and a corresponding wrong debug check in evaluate which
was protecting a node and then modifying it without unprotecting.

Also takes a step toward improving the unreadable cell to have a
canon heart byte and quote byte...but a much better idea for how to
lay out the node bits to avoid conflicting with UTF-8 strings came to
mind, so that will be implemented.