Pre-create all Lib builtin variables
This creates all the variables that the boot process calculated needed
symbols in Lib.  At the moment this isn't comprehensive (e.g. not all
mezzanine or base declarations), but it does cover at least the natives
and generics.  By doing the pre-creation, the Startup_Natives() and
Startup_Generics() routines can be simplified and run faster because
they don't have to pre-walk to create the declarations.