Reverse the sense of GC flags for Cells
This makes the flags for GC in cells indicate when *not* to mark the
cells, instead of when to mark them.

IT'S IN THE REVERSE SENSE ON PURPOSE.  This means a "free" cell can have
the following bit pattern WHICH IS NOT A LEGAL LEADING BYTE FOR UTF-8:

    11111xxx: Flags: NODE | FREE | GC_ONE | GC_TWO | CELL | ...

The free bit denotes an Init_Unreadable() cell, and so long as we set the
GC_ONE and GC_TWO flags we can still have free choices of `xxx` (e.g.
arbitrary ROOT, MANAGED, and MARKED flags), while Detect_Rebol_Pointer()
can be certain it's a cell and not UTF-8.

A couple more changes (like moving the cell bit) are needed before it
works fully, but it's close...