NODE_FLAG_MARKED to 0x01 for Detect_Rebol_Pointer()
This makes Detect_Rebol_Pointer() basically foolproof, in the sense
that if a pointer was ever a Stub* or a Cell* in its lifetime then
it should never conflate with UTF-8... *and* if it's in the freed
state then it should return as DETECTED_AS_FREE, with no conflation
with strings.

It also tightens up behaviors related to "decayed flexes", where
the Stub vs. Flex distinction is able make it so that Stub can be
decayed but Flex is not.