Make FLAVOR_0 out of band for flavor, Source subclass
There was a FLAVOR_CORRUPT, but this is better when we reserve FLAVOR_0
and then make that something that can be used for situations where you
want to specify a flavor that is optional.

Propagating this change led to several related changes, including
a subclass of array for Source specifically.  There were already some
weak checks based on the recognition that something like a Varlist
could not be used for a BLOCK! or GROUP! etc., but by separating it
completely it makes it so that the machinery for detecting if an
array has file/line is much simpler based on being the source
subclass.