Rethink TO Conversions
The original idea for TO and MAKE didn't separate them very well.  There
were few stakes in the ground for what should or should not count as a
TO conversion.

This commit formalizes TO as a non-evaluative operation that is strictly
reversible.  e.g. a TO conversion will only succeed if the reverse would
produce the same value back.  In addition to implementing that rule, it
also makes TO itself enforce this in the debug build by reversing the
transformation and asserting equivalence.