Exploit 0x00 initialization in Prep_Array()
Also makes Pop_Stack_Values_Core() use memcpy.  This gives it the
semantics of Move_Cell() with CELL_MASK_ALL, which was only used some
of the time...but it turns out to be the right thing to do in all
cases in order to get the memcpy() advantage in the release build.