Unreadable cells always unreadable, in release build too
Once a debug-only construct, unreadable cells are now foundational to the
system, and can be distinguished from UTF-8, stubs, or freed pool entries
in all circumstances.

However, there was still a residual #define for the idea that release
builds would not necessarily need to handle unreadable cells.  The cases
that depended on this were simply broken.

This removes that #define, though it points out that some more work needs
to be done in narrowing the contexts where unreadable cells can be
legitimately found in the release build.