Rethink DEBUG alternative to NDEBUG as "RUNTIME_CHECKS"
The NDEBUG macro was given a weird name by the standard to control
asserts, because DEBUG had been used for other purposes in code prior
to the standard.  So it was in the negative sense, making it awkward
to use with `#ifndef NDEBUG` being how you would test a "debug build"

A change was made at one point to translate this into a normal DEBUG
preprocessor define that would be either 1 or 0, so it could be
tested more reliably (since #if requires the argument to be defined,
so you avoid typos this way.)  However, it wasn't propagated fully
in the system due to an inkling that the term "DEBUG" was overloaded
for a build with runtime checks, as it conflated with features that
might be related to debugging code in the interpreter.

But there wasn't a resolved name for the alternative, which would
be called a "checked" build (similar to things like Windows NT that
use "checked" to mean built with instrumentation).  Rather than
`#if CHECKED` this goes with `#if RUNTIME_CHECKS`, and also defines
NO_RUNTIME_CHECKS.

Some minor issues were resolved while doing this update, and also a
notion of SPORADICALLY_NON_DETERMINISTIC() to randomly test when
tick-based testing isn't available (e.g. during startup).