Merge natives and generics, speed up downshifting
The desire to have common code that handles some preprocessing (or
postprocessing) on generically dispatched code makes it desirable to
just have "generics" be ordinary natives that do a delegation to
perform some amount of their work.

Simplifying in this way allows generics to be defined in the C code
(because they're just natives) and eliminates the %generics.r file
and all the extra handling related to that.  It's a big simplification.

The native entry point of the generic may want to do post-processing
(for instance, COPY wants to work on QUOTED!, but it wants to leverage
the underlying type's COPY behavior and then just add the incoming
quote level to that).  In order to do this gracefully in a way that
allows any given COPY generic to be implemented via continuations,
it's necessary to do this by using a new level.  The mechanic known
as "Push_Downshifted_Level()" was created for CASCADE, but it can be
used for generics...and this adds an accelerating twist to allow the
downshifted function to be called without returning to the trampoline.