Drop idea of const Flex being managed enforcement
There was an idea that by using C++ overloading, the system could catch
cases where `const String*` or `const Source*` were being passed in to
Init_Text() or Init_Block() or whatever and not being managed.  It could
not have any useful behavior, because the C build wouldn't be able to
do different things depending.  It could only give an assert.

It was somewhat obtuse, and also not even generically true.  Functions
that take (const Flex*) do so just to say they don't modify them.
In fact, it broke this:

    Init_Group(OUT, Freeze_Source_Shallow(a))

The Freeze routine gave back a const pointer.  But it didn't do anything
to the managed status.

No interesting bugs were ever caught with this check anyway, so it was
not earning its keep.