Consistent TAG! delimiting, FIND matches quoted molds
There was inconsistent behavior regarding whether TAG! was just
another ANY-STRING! with different delimiters, or if it had some
special behavior that made its delimiters show up.  This leans to
the simpler and more consistent model that the delimiters are
not there, but then compensates with a general feature that if
you use quoted things (e.g. in PARSE or FIND) when searching in
strings then you will match the molded content.  This provides
generic power:

    >> parse "<a> 100 (b c)" ['<a> space '100 space '(b c)]
    == (b c)