BINARY! => BLOB! conversion
The overloaded term "binary" carries a lot of connotations, including the
idea of just 1s and 0s #{110010100101} or that of an EXE program.  It
also conflated the internal storage type `Binary` (just a byte-sized
Flex, no positional information) with the Cell type that has an index.

Since JavaScript and other languages have gone with BLOB for binaries,
this adopts the shorter term.  It makes the source code clearer when
you are talking about a cell (BLOB!) vs. the storage (Binary), and is
overall an improvement.

This finally commits to it.