Introduce ENCODE 'UTF-8 and DECODE 'UTF-8, work in progress
This is another step in the gruelling task of trying to make TO
and MAKE and somewhat coherent.

For performance, the AS BLOB! should be used in some cases, though
it has the downside that the resulting BLOB! is constrained to UTF-8
modifications since there may still be string references to the
same series.