Check TO/COPY equivalence vs. implement w/COPY
As TO was being formalized the rule came up that (TO (TYPE OF X) X) would
have to be the same thing as COPY.

At first this was "enforced" by making the TO entry point notice when
the target type was the same as the input type, and call COPY instead of
the generic.  But now that there's a To_Checker_Executor(), it makes more
sense to instead call TO normally and then be able to do the check.
This tests that the code which copies naturally degrades into a TO, and
saves on paranoid asserts in TO that it wasn't called with the same
type as FROM.