Some PANIC cleanup, remove DIAGNOSE native
The Panic() code with diagnostics for DEBUG_STUB_ORIGINS hasn't had
to be used for a long time (which speaks well for the stability of
the system).

But it had atrophied, and the terminology wasn't up to date.  This
improves it a bit.  It also gets rid of the DIAGNOSE native, which
has never been used and would likely be of no value to anyone not
already running under a debugger.