Don't assume DEBUG_STUB_ORIGINS if DEBUG_COUNT_TICKS
A sensitive debugging scenario in the bootstrap executable revealed
that DEBUG_COUNT_TICKS would change the size of stubs, which was not
expected.  Sometimes a bug only manifests when the stub size is at
a particular amount, and if flipping the count ticks on and off
changes the stub size that's unexpected.

Separate the two completely.