Graft modern unreadable cell strategy onto old EXE
Unreadable cells originated as a debug-only concept, which were a
variant of BLANK! that would just act like a normal blank in release
builds.  It turned out that this unreadable non-state was actually
useful in more places, and also that a representation could be
chosen for it that used the second bit in the node header, yet still
pushed it out of band of characters that could start UTF-8.

This brings that modern understanding to the bootstrap executable,
and also cleans up how tracking information is added--abandoning
the edge case of tracking a few cells that didn't use their EXTRA
field (it's not that useful to warrant the divergence, if you need
the information use a DEBUG_TRACK_EXTEND_CELLS build).