Simplify Trampoline Level to not be global
The trampoline had its level stored in a longjmp list, which was an
artifact of some old pre-trampoline implementation methods.  In order
to make it easier to merge the behaviors of an abrupt failure vs.
throw or longjmp with what happens if you just `return FAIL()`, this
moves to a method that doesn't lose the current level when you drop
a rescue scope.

The results are simpler.  Since the "current" level (not necessarily
the top level) is stored in a local variable of the trampoline, it
need not be checked for disruption by other code.  It also makes the
variable available for easier inspection in the debugger.  This uses
the name "L" for the variable, which also makes the code a bit easier
on average to take in.