BOUNCE_FAIL Simplification for Abrupt Failures
It was a goal that executors and dispatchers should be able to write
`return FAIL()` and have the same handling path as `fail()`, where
the abrupt failure could be passed back to the executor or dispatcher
from which it originated so they could do cleanup.

The mechanism was more convoluted than it had to be, because a flag
for ABRUPT_FAILURE was on the level.  This creates a BOUNCE_FAIL that
is distinct from BOUNCE_THROWN, and that serves as the flag.  Then,
the dispatchers handle it normally with DISPATCHER_CATCHES.

This also simplifies some of the mechanics for DISPATCHER_CATCHES
by letting dispatchers set it themselves, instead of having it be
part of a continuation.  This means they can set up a catch state
even if they haven't requested a continuation (however they need to
get out of STATE_0 to indicate this was done consciously).