Eliminate ST_GROUP_BRANCH_ENTRY_DONT_ERASE_OUT
For some reason, the idea of slipping a parameter to the group
branch executor was done with putting it in the out cell, even
though the output cell gets erased by the continuation in state
zero... so a special state was used to tell the trampoline not
to erase it.  Easier is to use the SPARE cell.

Also, stops use of BOUNCE_DELEGATE by the group branch executor,
that is now an Action_Executor() feature only.