Fix use of non-designated LEVEL_FLAG_24
At one point, since LEVEL_FLAG_24 was free, it was used for a
miscellaneous purpose by ANY and ALL.  But then it became used
and signaled the wrong thing.  This allocates a named level flag
to help find such usages in the future: LEVEL_FLAG_MISCELLANEOUS,
now moved to bit position 23.