Re-enable error on #if of undefined things
Prior to the switch over of using #if instead of #ifdef systemically for
compilation flags, the Microsoft C warning for using #if on an undefined
thing was disabled, in order to be able to use /Wall.

It's better to be using #pragmas to disable the warning only for the
inclusion of windows.h, now that #if is used.