Give more direct API for parameter introspection
The concept of historical PARAMETERS OF is that it would give you the
decorated form of words--like in a function spec--and you would have
to re-process those to determine what they meant.  However, the way
things have been going the idea is to expose what the system knows
directly with properties of the parameters themselves.

Parameters don't have symbols (they describe the parameter, and then
sit in the named object slot).  So if you want the decorated form
(e.g. to show in help) then this gives you a function to apply that
decoration to any legal value.  This is used to an interesting effect
in the molding of parameters themselves, to put the decoration onto
the spec block and not a word (since the parameter has no symbol).

This touches upon some very broken frame and exemplar code that
needs to be rewritten pretty much top to bottom.  But these examples
are what motivates that redesign, so it's good to sketch it out.