_ is blank or space in parse, @ is synonym for ONE
There had been a bit of back and forth questioning the use of _
to match spaces in strings and literal blanks in lists, due to the
need to have a way to match single elements in a succinct way.

But blank for space is just too good, and it's not really that
satisfying as a "match this spot" abstraction.  Since @var and
@(expression) are ways of matching at the current location, the
idea that a plain @ would match anything (without the constraint
of var and expression) is actually relatively coherent.  And it's
a bit easier to see.

People who want more words and fewer symbols would likely choose
`ONE` or `<SKIP>` instead of @, but offering a short form is
something that the RegEx inclined would want to use, so it's kind
of important to have an answer for them.

This solidifies JUST as a means of synthesizing literal material
without matching it (no need for a group), and goes ahead and
declares LITERAL as the means of matching something literally, which
can help readability vs. adding another quoting level to something
that is already quoted.