Implement AS via Generic Dispatch
The AS operation had the same problem as TO in terms of not being a good
generic, because generics dispatched off of the first element in the
frame.  But when generics were changed to just be natives that called
the generic dispatcher, that could be changed on a native-by-native
basis.

(A better model is needed, but that just nudged the design a little in
the direction of more choice, and simpler code.)

This lets AS take advantage of the same code splitting benefits, so it
is much easier to follow.  It also allows conversions to proceed in
steps, e.g. a BINARY! blob which is being turned into a WORD! can
do so by doing a basic string conversion first, and then handing
that off to the string to word conversion.

Though it's still messy and buggy, it's a definite improvement.