Restore small UTF-8 immutable optimization
The optimization that ISSUE! stores its payload in its cell made the
unification of ISSUE! with CHAR! more palatable.  Somewhat haphazard
changes merging the EMAIL! and URL! types into the same class led
to cases of not using the in-cell optimization, because a string
was on hand...so the assert in the GC was dropped.  With the reshaping
of the AS function it became easier to refactor.